Alteracao na service e controller de produto e categoria, resolvido o bug de salvar ids fantasmas

+++ src/main/java/br/edu/ifsc/estoquecrud/service/ProdutoService.java
@@ -5,7 +5,9 @@ import br.edu.ifsc.estoquecrud.entity.Produto;
 import br.edu.ifsc.estoquecrud.repository.CategoriaRepository;
 import br.edu.ifsc.estoquecrud.repository.ProdutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.math.BigDecimal;
 import java.util.List;
@@ -117,10 +119,14 @@ public class ProdutoService {
     }
 
     //Deletar um produto
-    public String delete(Long id){
-        produtoRepository.deleteById(id);
+    public void delete(Long id){
 
-        return "Produto deletado com sucesso";
+        if (!produtoRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Produto não encontrado");
+        }
+
+        produtoRepository.deleteById(id);
     }
+
 
 }
